+++ src/main/java/duke/Duke.java
@@ -1,14 +1,6 @@
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeParseException;
+package duke;
+
 import java.util.Scanner;
-import java.util.ArrayList;
-import java.util.regex.Pattern;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 
 public class Duke {
     private Storage storage;
