+++ src/main/java/duke/DurationUnclearException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class DurationUnclearException extends Exception {
 }
