+++ src/main/java/duke/EmptyDescriptionException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class EmptyDescriptionException extends Exception {
 }
