+++ src/main/java/duke/InvalidTaskException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class InvalidTaskException extends Exception {
 }
