+++ src/main/java/duke/InvalidTextException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class InvalidTextException extends Exception {
 }
