+++ src/main/java/duke/Storage.java
@@ -1,0 +1,42 @@
+package duke;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.Scanner;
+
+public class Storage {
+    private File file;
+    String filePath;
+    public Storage(String filePath) {
+        this.filePath = filePath;
+        try {
+            this.file = new File(filePath);
+            Scanner scan = new Scanner(System.in);
+            if (!this.file.exists()) {
+                this.file.createNewFile();
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public File load() {
+        return this.file;
+    }
+
+    public void rewrite(String newText) {
+        try {
+            FileWriter fw;
+            Scanner sc = new Scanner(this.file);
+            fw = new FileWriter(filePath);
+            fw.write(newText);
+            fw.close();
+        } catch (FileNotFoundException e) {
+            System.out.println("File not found");
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
