+++ src/main/java/duke/Ui.java
@@ -1,13 +1,6 @@
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
+package duke;
+
 import java.util.Scanner;
-import java.util.regex.Pattern;
 
 public class Ui {
